Update convert.py
also added feature to not over ride duplicate chats with same name (add _v1, _v2, and so forth).
and also not not generate .md files with no name, which can happen for some reason, like ".md" (which obsidian will not show. instead it names those "noname" and "noname_v1" and so on.